+++ packages/uniforms-unstyled/__tests__/SubmitField.js
@@ -11,3 +11,19 @@ test('<SubmitField> - renders', () => {
 
     expect(wrapper).toHaveLength(1);
 });
+
+test('<SubmitField> - renders disabled if error', () => {
+    const element = <SubmitField />;
+    const wrapper = mount(element, createContext(undefined, {error: {}}));
+
+    expect(wrapper).toHaveLength(1);
+    expect(wrapper.find('input').prop('disabled')).toBe(true);
+});
+
+test('<SubmitField> - renders enabled if error and enabled', () => {
+    const element = <SubmitField disabled={false} />;
+    const wrapper = mount(element, createContext(undefined, {error: {}}));
+
+    expect(wrapper).toHaveLength(1);
+    expect(wrapper.find('input').prop('disabled')).toBe(false);
+});